array: don't inline array growth
Some versions of GCC cannot enforce strict aliasing rules when it is
inlined such that the size does not get updated for the calculation of
the offset for the new element.

Move this function to a different module so it does not get inlined and
the function is forced to load the values.